Revert "Feature/view model ut"
