+++ app/build.gradle
@@ -32,10 +32,6 @@ android {
     }
 
     buildFeatures.dataBinding = true
-
-    testOptions {
-        unitTests.returnDefaultValues = true
-    }
 }
 
 dependencies {
@@ -59,18 +55,18 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.1'
     implementation 'com.google.code.gson:gson:2.8.6'
 
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
+
     implementation 'org.koin:koin-core:2.2.1'
     implementation 'org.koin:koin-androidx-viewmodel:2.2.1'
 
     implementation 'com.squareup.picasso:picasso:2.71828'
 
-    testImplementation 'junit:junit:4.13.1'
+    testImplementation 'junit:junit:4.+'
     testImplementation 'org.koin:koin-test:2.2.1'
-    testImplementation "org.mockito:mockito-core:3.4.6"
     testImplementation 'com.nhaarman.mockitokotlin2:mockito-kotlin:2.0.0-alpha01'
     testImplementation "com.squareup.okhttp3:mockwebserver:4.9.1"
-    testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.3.9'
-    testImplementation "androidx.arch.core:core-testing:2.1.0"
 
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
