+++ app/src/main/java/br/com/franmoraes/githubrepositories/presentation/ui/repolist/GithubReposListActivity.kt
@@ -32,7 +32,6 @@ class GithubReposListActivity : AppCompatActivity() {
 
         prepareObservers()
         initViews()
-        viewModel.fetchRepositories()
     }
 
     private fun prepareObservers() {
