--- app/src/test/java/br/com/franmoraes/githubrepositories/base/CoroutineTestRule.kt
@@ -1,31 +1,0 @@
-package br.com.franmoraes.githubrepositories.base
-
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.*
-import org.junit.rules.TestRule
-import org.junit.runner.Description
-import org.junit.runners.model.Statement
-
-
-@ExperimentalCoroutinesApi
-class CoroutineTestRule : TestRule {
-
-    private val testDispatcher = TestCoroutineDispatcher()
-    private val testScope = TestCoroutineScope(testDispatcher)
-
-    override fun apply(base: Statement, description: Description?) = object : Statement() {
-        @Throws(Throwable::class)
-        override fun evaluate() {
-            Dispatchers.setMain(testDispatcher)
-
-            base.evaluate()
-
-            Dispatchers.resetMain()
-            testScope.cleanupTestCoroutines()
-        }
-    }
-
-    fun runTestBlock(block: suspend TestCoroutineScope.() -> Unit) =
-        testScope.runBlockingTest { block() }
-}
